refactor: combine add item and list forms
