+++ domus-frontend/src/app/listes/add-form/add-form.spec.ts
@@ -1,0 +1,21 @@
+import {ComponentFixture, TestBed} from '@angular/core/testing';
+import {AddForm} from './add-form';
+
+describe('AddForm', () => {
+  let component: AddForm;
+  let fixture: ComponentFixture<AddForm>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [AddForm]
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(AddForm);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
