--- domus-frontend/src/app/listes/add-item-form/add-item-form.html
@@ -1,16 +1,0 @@
-<form #itemForm="ngForm" (ngSubmit)="addItem(itemForm)">
-  <input type="text"
-         id="item-input"
-         name="item"
-         [(ngModel)]="newItemText"
-         (input)="searchSuggestions()"
-         (change)="maybeAddSuggestion()"
-         list="item-suggestions"
-         placeholder="Ajouter un item">
-  <datalist id="item-suggestions">
-    @for (s of suggestions; track s.id) {
-      <option [value]="s.text"></option>
-    }
-  </datalist>
-  <button type="submit">Ajouter</button>
-</form>
