--- domus-frontend/src/app/listes/add-list-form/add-list-form.html
@@ -1,5 +1,0 @@
-<form #addForm="ngForm" (submit)="addList(addForm);">
-  <input type="text" [(ngModel)]="newName" name="name" placeholder="Nom de la liste">
-  <input type="text" [(ngModel)]="newTag" name="tag" placeholder="Tag (ex: course, été)">
-  <button type="submit">Ajouter</button>
-</form>
