--- domus-frontend/src/app/listes/add-list-form/add-list-form.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { AddListForm } from './add-list-form';
-
-describe('AddListForm', () => {
-  let component: AddListForm;
-  let fixture: ComponentFixture<AddListForm>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [AddListForm]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(AddListForm);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
